+++ rnalysis/data_files/r_templates/limma_lrt_parametric.R
@@ -1,5 +1,5 @@
 
 bayes_fit <- eBayes(fit)
 lrt_res <- topTable(bayes_fit, n=Inf, coef=$COEFS)
-lrt_res_ordered <- lrt_res[order(res$adj.P.Val),]
+lrt_res_ordered <- lrt_res[order(lrt_res$adj.P.Val),]
 write.csv(as.data.frame(lrt_res_ordered),file="$OUTFILE_NAME")