+++ rnalysis/data_files/r_templates/limma_run_parametric.R
@@ -7,6 +7,6 @@ colnames(design)[1] <- "Intercept" # rename the intercept column from "(Intercep
 colnames(design) <- make.names(colnames(design)) # make coefficient names syntactically correct
 
 count_data <- read.table("$COUNT_MATRIX", header=TRUE, sep= ",", row.names = 1)
-voom_object <- voom(count_data, design, plot=FALSE, save.plot=TRUE)
+voom_object <- $VOOMFUNC(count_data, design, plot=FALSE, save.plot=TRUE)
 
 $RANDOM_EFFECT_FIT