+++ tests/test_files/limma_tests/case3/expected_limma_script_3.R
@@ -4,16 +4,16 @@ design_matrix <- read.table("tests/test_files/test_design_matrix.csv", header=TR
 design_matrix$condition <- factor(design_matrix$condition, levels=c("cond1", "cond2", "cond3"))
 design_matrix$replicate <- factor(design_matrix$replicate, levels=c("rep1", "rep2", "rep3"))
 
-design <- model.matrix(~ condition + replicate, design_matrix)
+design <- model.matrix(~ condition, design_matrix)
 colnames(design)[1] <- "Intercept" # rename the intercept column from "(Intercept)" to "Intercept"
 colnames(design) <- make.names(colnames(design)) # make coefficient names syntactically correct
 
 count_data <- read.table("tests/test_files/big_counted.csv", header=TRUE, sep= ",", row.names = 1)
 voom_object <- voom(count_data, design, plot=FALSE, save.plot=TRUE)
 
-cor <- duplicateCorrelation(voom_object, design, block=replicate)
+cor <- duplicateCorrelation(voom_object, design, block=design_matrix$replicate)
 if (cor$consensus.correlation > 0) { #only include random effect if the correlation is positive
-  fit <- lmFit(voom_object, design, block=replicate, correlation=cor$consensus.correlation)
+  fit <- lmFit(voom_object, design, block=design_matrix$replicate, correlation=cor$consensus.correlation)
 }else {   fit <- lmFit(voom_object, design)}
 
 contrast <- makeContrasts("conditioncond2", levels = design)